Favorites routes
Add a profile Favorite route, and update list route so you can add items to lists.

+++ Api/src/routes/users.ts
@@ -18,7 +18,7 @@ const upload = multer({
   },
 });
 import axios from "axios";
-import { deleteToTheList } from "../utils/lists";
+import { deleteToTheList, addToTheList } from "../utils/lists";
 export const usersRouter = Router();
 
 usersRouter.get("/", async (req, res) => {
@@ -375,12 +375,19 @@ usersRouter.put<{id: string, list: string}, {}>("/user/lists/:id", isAuthenticat
   const { mangaId } = req.body;
   
   if (list !== "library" && list !== "favorites" && list !== "wishList" ) return res.status(400).send({msg: "Invalid list name"});
-      
+  
   try {
+    //@ts-ignore
+    if( req.user[list].includes(number(mangaId))){
     let mangasList = await deleteToTheList(id, list, mangaId);
     
     return (mangasList.length === 0) ?
     res.send({msg: "Empty list"}) : res.send({msg: "Delete manga to the list"})
+    } else {
+      let user = await addToTheList(id, list, mangaId);
+      return (user[list].length === 0) ?
+      res.send({msg: "Empty list"}) : res.send({msg: "Add manga to the list"})
+    }
   } catch (error: any) {
     console.log("Delete manga from list: ", error)
     return res.status(400).send({error: error.message})
